% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add or Edit Questions.R
\name{add_mc}
\alias{add_mc}
\title{Add Multiple Choice Questions to an Existing Quiz}
\usage{
add_mc(test, ans, qus, wts = rep(1, qus), ats = rep(1, qus))
}
\arguments{
\item{test}{The test object that you would like to add questions too}

\item{ans}{The number of answers for each new question}

\item{qus}{The number of questions to be added to the quiz}

\item{wts}{The weighting score for each question, can be provided as a single integer, or a vector with number of elements equal to \code{qus}}

\item{ats}{The number of attempts the test taker is allowed to answer the question, can be provided as a single integer, or a vector with number of elements equal to \code{qus}.}
}
\value{
A new test object, with the new questions appended onto the end
}
\description{
Add Multiple Choice Questions to an Existing Quiz
}
\examples{
add_mc(test1,4,10)
add_mc(test1,c(2,2,3,3))
add_mc(test1,4,5,1:5,2)
}
