% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add or Edit Questions.R
\name{add_tf}
\alias{add_tf}
\title{Add True/False Questions to an Existing Quiz}
\usage{
add_tf(test, qus, wts = rep(1, qus), ats = rep(1, qus))
}
\arguments{
\item{test}{The test object that you would like to add questions too}

\item{qus}{The number of questions you would like to add}

\item{wts}{The weighting score for each question, can be provided as a single integer, or a vector with number of elements equal to \code{qus}}

\item{ats}{The number of attempts the test taker is allowed to answer the question, can be provided as a single integer, or a vector with number of elements equal to \code{qus}. For any value greater than 1 the answer will be correct every time.}
}
\value{
A new test object, with the new questions appended onto the end
}
\description{
Add True/False Questions to an Existing Quiz
}
\examples{
add_tf(test1,10)
add_tf(test1,5,1:5)
}
