% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate Results.R
\name{below_percentage}
\alias{below_percentage}
\title{Calculate the portion of scores that are lower than a given percentage score}
\usage{
below_percentage(result, below, inclusive = TRUE)
}
\arguments{
\item{result}{The result object}

\item{below}{The maximum score for the desired range}

\item{inclusive}{Logical value, whether scores equal to the maximum score should be included}
}
\value{
A percentage score showing the percentage of scores than were lower than the pass score
}
\description{
Calculate the portion of scores that are lower than a given percentage score
}
\examples{
below_percentage(result1,0.8)
below_percentage(result1,0.7,FALSE)
}
