% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate Results.R
\name{between_percentage}
\alias{between_percentage}
\title{Calculate the portion of scores that are between a given set of percentage scores}
\usage{
between_percentage(result, low, high, inclusive = TRUE)
}
\arguments{
\item{result}{The result object}

\item{low}{The 'pass' score, i.e. the minimum score for the desired range}

\item{high}{The maximum score for the desired range}

\item{inclusive}{Logical value, whether scores equal to the low and high score should be included}
}
\value{
A percentage score showing the percentage of scores than were between the two provided scores
}
\description{
Calculate the portion of scores that are between a given set of percentage scores
}
\examples{
between_percentage(result1,0.4,0.8)
between_percentage(result1,0.5,0.7,FALSE)
}
