% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiple Answers.R
\name{create_maq}
\alias{create_maq}
\title{Create a Multiple Choice Quiz with Multiple Answers Required}
\usage{
create_maq(ans, qus, cor = rep(1, qus), wts = rep(1, qus), ats = rep(1,
  qus))
}
\arguments{
\item{ans}{The number of answers for each question, can be a single integer, or a vector providing the number of answers for multiple questions}

\item{qus}{The number of questions in the quiz, not neccesary if a vector is provided for \code{ans}}

\item{cor}{The number of answers that must be given, for questions where, for example, 2 of 4 answers are correct.}

\item{wts}{The weighting score for each question, can be provided as a single integer, or a vector with number of elements equal to \code{qus}}

\item{ats}{The number of attempts the test taker is allowed to answer the question, can be provided as a single integer, or a vector with number of elements equal to \code{qus}}
}
\value{
A test object
}
\description{
Create a Multiple Choice Quiz with Multiple Answers Required
}
\examples{
create_maq(4,10,2)
create_maq(c(4,4,5,5),3)
create_maq(4,5,1:5,3,2)
}
