% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_questions.R
\name{create_mc}
\alias{create_mc}
\title{Create a Multiple Choice Quiz}
\usage{
create_mc(ans, qus = length(ans), wts = rep(1, qus), ats = rep(1, qus))
}
\arguments{
\item{ans}{The number of answers for each question, can be a single integer, or a vector providing the number of answers for multiple questions}

\item{qus}{The number of questions in the quiz, not neccesary if a vector is provided for \code{ans}}

\item{wts}{The weighting score for each question, can be provided as a single integer, or a vector with number of elements equal to \code{qus}}

\item{ats}{The number of attempts the test taker is allowed to answer the question, can be provided as a single integer, or a vector with number of elements equal to \code{qus}}
}
\value{
A test object
}
\description{
Create a Multiple Choice Quiz
}
\examples{
create_mc(4,10)
create_mc(c(2,2,3,3))
create_mc(4,5,1:5,2)
}
