% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_questions.R
\name{create_tf}
\alias{create_tf}
\title{Create a True/False Quiz}
\usage{
create_tf(qus, wts = rep(1, qus), ats = rep(1, qus))
}
\arguments{
\item{qus}{The number of questions in the quiz}

\item{wts}{The weighting score for each question, can be provided as a single integer, or a vector with number of elements equal to qus}

\item{ats}{The number of attempts the test taker is allowed to answer the question, can be provided as a single integer, or a vector with number of elements equal to \code{qus}. For any value greater than 1 the answer will be correct every time.}
}
\value{
A test object
}
\description{
Create a True/False Quiz
}
\examples{
create_tf(10)
create_tf(5,1:5)
}
