% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate Results.R
\name{mean.result}
\alias{mean.result}
\title{Calculate the mean score of the results}
\usage{
\method{mean}{result}(result)
}
\arguments{
\item{result}{The result object}
}
\value{
The mean score
}
\description{
Calculate the mean score of the results
}
\examples{
mean.result(result1)
}
