% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate Results.R
\name{mean_result}
\alias{mean_result}
\title{Calculate the mean score of the results}
\usage{
mean_result(result)
}
\arguments{
\item{result}{The result object}
}
\value{
The mean score
}
\description{
Calculate the mean score of the results
}
\examples{
mean_result(result1)
}
