% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate Results.R
\name{median.result}
\alias{median.result}
\title{Calculate the median score of the results}
\usage{
\method{median}{result}(result)
}
\arguments{
\item{result}{The result object}
}
\value{
The median score
}
\description{
Calculate the median score of the results
}
\examples{
median.result(result1)
}
