% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate Results.R
\name{median_result}
\alias{median_result}
\title{Calculate the median score of the results}
\usage{
median_result(result)
}
\arguments{
\item{result}{The result object}
}
\value{
The median score
}
\description{
Calculate the median score of the results
}
\examples{
median_result(result1)
}
