% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate Results.R
\name{modal.result}
\alias{modal.result}
\title{Calculate the mode score of the results}
\usage{
\method{modal}{result}(result)
}
\arguments{
\item{result}{The result object}
}
\value{
The mode score
}
\description{
Calculate the mode score of the results
}
\examples{
modal.result(result1)
}
