% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate Results.R
\name{mode_result}
\alias{mode_result}
\title{Calculate the mode score of the results}
\usage{
mode_result(result)
}
\arguments{
\item{result}{The result object}
}
\value{
The mode score
}
\description{
Calculate the mode score of the results
}
\examples{
mode_result(result1)
}
