% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate Results.R
\name{pass_percentage}
\alias{pass_percentage}
\title{Calculate the portion of scores that are higher than a given percentage score}
\usage{
pass_percentage(result, pass, inclusive = TRUE)
}
\arguments{
\item{result}{The result object}

\item{pass}{The 'pass' score, i.e. the minimum score for the desired range}

\item{inclusive}{Logical value, whether scores equal to the pass score should be included}
}
\value{
A percentage score showing the percentage of scores than were greater than the pass score
}
\description{
Calculate the portion of scores that are higher than a given percentage score
}
\examples{
pass_percentage(result1,0.4)
pass_percentage(result1,0.5,FALSE)
}
