% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph Results.R
\name{plot.result}
\alias{plot.result}
\title{Plot a histogram of the results}
\usage{
\method{plot}{result}(result)
}
\arguments{
\item{result}{The result object}
}
\value{
A histogram showing the distribution of each possible score between 0 (no correct answers) and 1 (all correct answers)
}
\description{
Plot a histogram of the results
}
\examples{
plot.result(result1)
}
