% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run Sample.R
\name{sample_test}
\alias{sample_test}
\title{Run a simulation to create a sample of test scores}
\usage{
sample_test(test, trials, know = 0)
}
\arguments{
\item{test}{The test object}

\item{trials}{The desired number of iterations, i.e. the sample size}

\item{know}{The 'knowledge' of the test taker, i.e. a knowledge of 0.5 indicates that there is 50% chance they will know (meaning provide the correct) answer}
}
\value{
A result object
}
\description{
Run a simulation to create a sample of test scores
}
\examples{
sample_test(test1,100)
sample_test(test1,5000,0.25)
}
