% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate Results.R
\name{table.result}
\alias{table.result}
\title{Create a table of the results}
\usage{
table.result(result)
}
\arguments{
\item{result}{the result object}
}
\value{
A table showing percentage frequencies for each score
}
\description{
Create a table of the results
}
\examples{
table.result(result1)
}
