% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate Results.R
\name{var.result}
\alias{var.result}
\title{Calculate the variance of the scores}
\usage{
var.result(result)
}
\arguments{
\item{result}{The result object}
}
\value{
The variance of the scores
}
\description{
Calculate the variance of the scores
}
\examples{
var.result(result1)
}
