% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate Results.R
\name{variance.result}
\alias{variance.result}
\title{Calculate the variance of the scores}
\usage{
\method{variance}{result}(result)
}
\arguments{
\item{result}{The result object}
}
\value{
The variance of the scores
}
\description{
Calculate the variance of the scores
}
\examples{
variance.result(result1)
}
